chore: renamed removeUrlJestSerializer to removeRandomizedUrlJestSerializer
